feat: add ScreenScale plugin and refactor scale management
1. Add new ScreenScale plugin (org.deepin.dde.ScreenScale1) as the
single source of truth for screen scale factors
2. Implement GetScreenScaleInfo and SetScaleFactor DBus methods with
automatic recommended scale calculation
3. Refactor XSettings to delegate scale management to ScreenScale
service via DBus
4. Remove deprecated ScaleFactorHelper, updateFirefoxDPI, and related
legacy code
5. Standardize CMake variable naming from BIN_NAME to PLUGIN_NAME across
all plugins
6. Update copyright years from 2025 to 2025-2026
7. Use dtk_add_config_meta_files for DConfig installation

Log: Added centralized screen scale management service with automatic
recommended scale calculation based on monitor physical dimensions

Influence:
1. Test GetScreenScaleInfo with various screen configurations (single
monitor, multi-monitor, HiDPI)
2. Verify SetScaleFactor validates input range (1.0-3.0) and persists
via DConfig
3. Test scale factor synchronization between ScreenScale and XSettings
services
4. Verify ScaleFactorChanged signal propagation across services
5. Test recommended scale calculation accuracy for different monitor
sizes
6. Verify backward compatibility with existing scale configurations
7. Test plugin loading/unloading via deepin-service-manager
8. Verify DBus permission policies for ScreenScale1 service

feat: 添加屏幕缩放插件并重构缩放管理

1. 新增 ScreenScale 插件 (org.deepin.dde.ScreenScale1) 作为屏幕缩放比例
的唯一事实来源
2. 实现 GetScreenScaleInfo 和 SetScaleFactor DBus 方法，支持自动计算推荐
缩放比例
3. 重构 XSettings 插件，通过 DBus 将缩放管理委托给 ScreenScale 服务
4. 移除已废弃的 ScaleFactorHelper、updateFirefoxDPI 及相关遗留代码
5. 统一各插件 CMake 变量命名，将 BIN_NAME 改为 PLUGIN_NAME
6. 更新版权年份从 2025 到 2025-2026
7. 使用 dtk_add_config_meta_files 安装 DConfig 配置

Log: 新增屏幕缩放管理服务，支持根据显示器物理尺寸自动计算推荐缩放比例

Influence:
1. 测试 GetScreenScaleInfo 接口，覆盖单屏、多屏、HiDPI 等场景
2. 验证 SetScaleFactor 对输入范围 (1.0-3.0) 的校验及 DConfig 持久化
3. 测试 ScreenScale 与 XSettings 服务间的缩放比例同步
4. 验证 ScaleFactorChanged 信号在各服务间的传播
5. 测试不同显示器尺寸下推荐缩放比例计算的准确性
6. 验证与现有缩放配置的向后兼容性
7. 测试通过 deepin-service-manager 加载/卸载插件
8. 验证 ScreenScale1 服务的 DBus 权限策略

PMS: BUG-355227

+++ src/plugin-qt/xsettings/impl/xsdatainfo.cpp
@@ -1,4 +1,4 @@
-// SPDX-FileCopyrightText: 2025 UnionTech Software Technology Co., Ltd.
+// SPDX-FileCopyrightText: 2025 - 2026 UnionTech Software Technology Co., Ltd.
 //
 // SPDX-License-Identifier: LGPL-3.0-or-later
 #include "xsdatainfo.h"
@@ -46,7 +46,7 @@ void XSItemInfo::unMarshalXSItemInfoData(QByteArray &datas)
         value = tempString;
         break;
     case HeadTypeColor:
-        for (int i = 0; i < colorSize; i++) {
+        for (size_t i = 0; i < colorSize; i++) {
             Utils::readInteger(datas, tempColor[i]);
         }
         value = tempColor;
@@ -127,7 +127,7 @@ bool XSItemInfo::marshalXSItemInfoData(QByteArray &datas)
         if (colorValue == nullptr) {
             break;
         }
-        for (int i = 0; i < colorSize; i++) {
+        for (size_t i = 0; i < colorSize; i++) {
             Utils::writeInteger(datas, (*colorValue)[i]);
         }
         ret = true;
